@@ -1,20 +1,26 @@
 module github.com/ooaklee/ghatd
 
-go 1.22.0
+go 1.23.0
+
+toolchain go1.23.8
 
 require (
-	//>ghatd {{ block .DetailModGhatdPackage }}
-	github.com/PaesslerAG/jsonpath v0.1.1
 	//>ghatd {{ end }}
+	//>ghatd {{ block .DetailModGhatdPackage }}
 	github.com/NYTimes/gziphandler v1.1.1
+	github.com/PaesslerAG/jsonpath v0.1.1
+	github.com/ettle/strcase v0.2.0
 	github.com/ooaklee/http-cache v0.0.0-20240308024722-18826df341f3
+	github.com/ritwickdey/querydecoder v1.2.0
 	github.com/spf13/cobra v1.8.0
+	github.com/xakep666/mongo-migrate v0.2.1
 //>ghatd {{ block .WebDetailGoModRequirePackages }}{{ end }}
 //>ghatd {{ block .ApiDetailGoModRequirePackages }}{{ end }}
 )
 
 require (
-	github.com/go-git/go-git/v5 v5.11.0
+	github.com/go-git/go-git/v5 v5.14.0
+	github.com/mergestat/timediff v0.0.3
 	github.com/otiai10/copy v1.14.0
 )
 
