@@ -5,7 +5,10 @@ import (
 	"errors"
 	"math"
 
+	"github.com/ettle/strcase"
+	"github.com/ooaklee/ghatd/external/logger"
 	"github.com/ooaklee/reply"
+	"go.uber.org/zap"
 )
 
 const (
@@ -18,62 +21,93 @@ var GetResourcePaginationErrorMap = map[string]reply.ErrorManifestItem{
 	ErrKeyPageOutOfRange: {Title: "Bad Request", Detail: "Page out of range", StatusCode: 400},
 }
 
-// GetResourcePaginationResponse is the data shaped on the requested
-// item
-type GetResourcePaginationResponse struct {
+// ResponseMetaKey is a string type used as the keys in the map returned
+// by requests
+type ResponseMetaKey string
 
-	// Resources is the collection of the resource to paginate
-	Resources []interface{}
+// ResponseMetaKey is a string type used as the keys in the map returned
+// by requests
+const (
 
-	// Total - number of resources found
-	Total int
+	// ResponseMetaKeyResourcePerPage is the key for the number of resources per page
+	ResponseMetaKeyResourcePerPage ResponseMetaKey = "resources_per_page"
 
-	// TotalPages pages available
-	TotalPages int
+	// ResponseMetaKeyTotalResources is the key for the total number of resources
+	ResponseMetaKeyTotalResources ResponseMetaKey = "total_resources"
 
-	// ResourcePerPage is how many many resources
-	// are in the page
-	ResourcePerPage int
+	// ResponseMetaKeyTotalPages is the key for the total number of pages
+	ResponseMetaKeyTotalPages ResponseMetaKey = "total_pages"
 
-	// Page specifies the page results should be taken from. Default 1.
-	Page int
+	// ResponseMetaKeyPage is the key for the current page number
+	ResponseMetaKeyPage ResponseMetaKey = "page"
+)
+
+// PaginationRequest represents a standard pagination request
+type PaginationRequest struct {
+
+	// PerPage is the number of resources per page
+	PerPage int `json:"per_page"`
+
+	// Page is the current page number
+	Page int `json:"page"`
 }
 
-// GetResourcePaginationRequest is parameters to shape the data
-type GetResourcePaginationRequest struct {
-
-	// Total number of resources to return per page
-	PerPage int
-
-	// Page specifies the page results should be taken from. Default 1.
-	Page int
+// PaginationResponse represents a standard pagination response
+type PaginationResponse[T any] struct {
+	Resources       []T `json:"resources"`
+	Total           int `json:"total"`
+	TotalPages      int `json:"total_pages"`
+	ResourcePerPage int `json:"resource_per_page"`
+	Page            int `json:"page"`
 }
 
-// GetResourcePagination returns appropiate pagination response based on request and
-// passed resource slice
-func GetResourcePagination(ctx context.Context, r *GetResourcePaginationRequest, passedResources []interface{}, totalResources int) (*GetResourcePaginationResponse, error) {
-	var err error
+// Paginate handles pagination for any resource type
+func Paginate[T any](
+	ctx context.Context,
+	req *PaginationRequest,
+	resources []T,
+	totalCount int,
+) (*PaginationResponse[T], error) {
+	logger := logger.AcquireFrom(ctx).WithOptions(
+		zap.AddStacktrace(zap.DPanicLevel),
+	)
 
-	totalPage := int(math.Ceil(float64(totalResources) / float64(r.PerPage)))
+	// Set default values if needed
+	perPage := req.PerPage
+	if perPage <= 0 {
+		perPage = 10 // Default items per page
+	}
 
-	// If no resources in collection return empty list
-	if totalResources == 0 && r.Page == 1 {
-		return &GetResourcePaginationResponse{
-			Total:           totalResources,
-			TotalPages:      1,
-			Resources:       passedResources,
-			Page:            r.Page,
-			ResourcePerPage: r.PerPage,
+	page := req.Page
+	if page <= 0 {
+		page = 1 // Default to first page
+	}
+
+	// Calculate pagination parameters
+	totalPages := int(math.Ceil(float64(totalCount) / float64(perPage)))
+
+	// If no resources in collection
+	if totalCount == 0 && page == 1 {
+		return &PaginationResponse[T]{
+			Resources:       []T{},
+			Total:           totalCount,
+			TotalPages:      totalPages,
+			ResourcePerPage: perPage,
+			Page:            page,
 		}, nil
-	} else if r.Page > totalPage {
+	} else if page > totalPages {
+		logger.Warn("pagination-page-exceeds-total-pages",
+			zap.Int(string(ResponseMetaKeyPage), page),
+			zap.Int(strcase.ToKebab(string(ResponseMetaKeyTotalPages)), totalPages),
+		)
 		return nil, errors.New(ErrKeyPageOutOfRange)
 	}
 
-	return &GetResourcePaginationResponse{
-		Total:           totalResources,
-		TotalPages:      totalPage,
-		Resources:       passedResources,
-		Page:            r.Page,
-		ResourcePerPage: r.PerPage,
-	}, err
+	return &PaginationResponse[T]{
+		Resources:       resources,
+		Total:           totalCount,
+		TotalPages:      totalPages,
+		ResourcePerPage: perPage,
+		Page:            page,
+	}, nil
 }
