@@ -10,18 +10,21 @@ type Settings struct {
 	LogLevel              string `envconfig:"log_level" default:"info"`
 	Host                  string `default:"0.0.0.0"`
 	Port                  string `default:"4000"`
-	// Cache
+
+	// CACHE
 	CacheTtl                 int    `envconfig:"cache_ttl" default:"15" required:"true"`
 	CacheRefreshParameterKey string `envconfig:"cache_refresh_key" default:"frais" required:"true"`
 	CacheSkipHttpHeader      string `envconfig:"cache_skip_http_header" default:"x-cache-skip"`
 	CacheSkipUriPathRegex    string `envconfig:"cache_skip_uri_path_regex" default:"^/api/v1/.*"`
-	// Cors
+
+	// CORS
 	AllowOrigins string `envconfig:"allow_origins" default:"http://localhost:3000" required:"true"`
-	// Web App
-	ExternalServiceName    string `envconfig:"external_service_name" default:"GHAT(D)" required:"true"`
-	ExternalServiceWebsite string `envconfig:"external_service_website"  default:"https://ghatd.com" required:"true"`
-	ExternalServiceEmail   string `envconfig:"external_service_email"  default:"leon+ghatd@boasi.io" required:"true"`
-	LegalBusinessName      string `envconfig:"external_legal_business_name"  default:"GHAT(D)" required:"true"`
+
+	// ENTITY
+	BusinessEntityName      string `envconfig:"business_entity_name" default:"GHAT(D)" required:"true"`
+	BusinessEntityWebsite   string `envconfig:"business_entity_website" default:"https://ghatd.com"`
+	BusinessEntityEmail     string `envconfig:"business_entity_email" default:"leon+ghatd@boasi.io"`
+	BusinessEntityNameLegal string `envconfig:"business_entity_name_legal" default:"GHAT(D)" required:"true"`
 }
 
 // NewSettings returns app settings
